oh hey I already built a connection pool, let's use it
This commit was sponsored by Thomas Ballinger, Steven S., Sergio Bost,
and my other patrons.  If you want to join them, you can support my
work at https://glyph.im/patrons/.
